% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.}
}
\value{
The function returns an object of class linreg as an S3 class
}
\description{
Implementing the calculations for multiple regression model using QR decomposition.
}
\references{
http://genomicsclass.github.io/book/pages/qr_and_regression.html
}
