% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{This contains the plot methods for linereg function.}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of linereg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the plot methods for linereg function.
}
