% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{This contains the fitted values of linreg function.}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object of linreg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the fitted values of linreg function.
}
