% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{This contains the print methods for linreg function.}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of linreg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the print methods for linreg function.
}
