% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{resid.linreg}
\alias{resid.linreg}
\title{This contains the residuals of linreg function.}
\usage{
\method{resid}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of linreg class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
This contains the residuals of linreg function.
}
